% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dir_rbind}
\alias{dir_rbind}
\title{Rowise combine all the files in a directory}
\usage{
dir_rbind(dir_path, header = F, col_names = NULL, out_file_path = NULL)
}
\arguments{
\item{dir_path}{Absolute path of the directory containing the files to be combined.}

\item{header}{Boolean value of if the header needs to be considered}

\item{col_names}{Character vector to replace the header with a custom column name.}

\item{Outputl}{File path - absolute path. By default its null and can be omitted. Produces a tsv file.}
}
\description{
This function combines the files in the given directory and returns as a data table or saves it to the out_put_file_path.
Option 1 : Set header = T and no col_names. In this case the function will read the file and use the header of the first 
file in the directory. 
Option 2 : Set header = F and specify col_names manually. In this case the function will read the file and will use the given array as column names
The assumption is that it does not have a header (default)
}

