% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-snp-ids.r
\name{execute_script}
\alias{execute_script}
\title{execute_script() the driver function for the script}
\usage{
execute_script(in_csv_file_path, in_ref_tsv_file_path,
  out_snp_name_annotated_study_snps_file_path, in_ref_snps_dir,
  out_combine_ref_snp_tsv_file_path)
}
\arguments{
\item{csv_file_path}{A csv file (absolute)path  with data about snps and stats from 
different studies}

\item{ref_tsv_file_path}{A tsv refernce file (absolute)path}
}
\description{
execute_script() the driver function for the script
}

