% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/get-snp-ids.r
\name{generate_new_ids}
\alias{generate_new_ids}
\title{The function takes in the data frame after attaching snp ids from the ref.
using get_snp_ids(). The function computes the length of missing snp_name}
\usage{
generate_new_ids(df)
}
\arguments{
\item{the}{dataframe df}
}
\value{
a dataframe with all missing snp names changed to pgi_dal_snp1,...
}
\description{
The function takes in the data frame after attaching snp ids from the ref.
using get_snp_ids(). The function computes the length of missing snp_name
}

