% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant-snps.r
\name{get_max_and_mean}
\alias{get_max_and_mean}
\title{calculates snp count, max and mean on given column name and groups all the counts by chromosome
if the out file path is not given then it returns the datatable else writes a tsv on the new path}
\usage{
get_max_and_mean(df, column_name, out_file_path = NULL)
}
\description{
calculates snp count, max and mean on given column name and groups all the counts by chromosome
if the out file path is not given then it returns the datatable else writes a tsv on the new path
}

