% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant-snps.r
\name{get_nlp}
\alias{get_nlp}
\title{if the out file path is not given then it returns the datatable else writes a tsv on the new path}
\usage{
get_nlp(df, column_name, out_file_path = NULL)
}
\description{
if the out file path is not given then it returns the datatable else writes a tsv on the new path
}

