% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant-snps.r
\name{get_topQ}
\alias{get_topQ}
\title{Top Q statistics.
Uses Hmisc cut2 to get mean of nlp(negative log p-value) of snps in the top quartile of each gene}
\usage{
get_topQ(df, column_name, threshold = 25, out_file_path = NULL)
}
\description{
@param df data.table with the snps maps to genes
 @param column_name the name of the column which have negative log p-values
 @param quartile the quartile you want to use for topQ stats
 @param out_file_path The file where  output will be writen to
Note: Internally calls another function get_quartile
}

