% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/get-snp-ids.r
\name{main}
\alias{main}
\title{main() the driver function for the script}
\usage{
main(filename1)
}
\arguments{
\item{filename1}{A csv file path from / with data aboout snps and stats from different studies}
}
\description{
main() the driver function for the script
}

