% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps-to-genes.r
\name{map_snps_to_gene}
\alias{map_snps_to_gene}
\title{It finds all the snps in genome that are in  gene +/- window_size}
\usage{
map_snps_to_gene(study_df, ref_df, window_size = 1000, out_file_path = NULL)
}
\arguments{
\item{study_df}{: a datatable with snp information identified in the study. Should have column names - chr_no, snp_pos}

\item{ref_df}{: is the Bos taraus data which has gene_start, gene_end and chr_no column names}

\item{window_size}{: gene start and end +/- window_size}
}
\value{
a dt with all snps in gene windows
}
\description{
It finds all the snps in genome that are in  gene +/- window_size
}

