% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-correction.r
\name{p_adjustment}
\alias{p_adjustment}
\title{This function add a new column to the datatable created by bonferroni correction
should have a column 'chr' which is integers only}
\usage{
p_adjustment(study_df, col_names, level = "genome", test = "fdr",
  out_file_path = NULL)
}
\arguments{
\item{col_names}{Vector of columns to be acted upon}

\item{out_file_path}{Optional parameter. The file path to store the output.}

\item{in_un_adj_p_val_snps_data_file_path}{is tab separated filepath. This is the file with 
unadjusted p valu columns.}
}
\value{
DataTable with additional columns for genomewide correction of each column vector



# p_adjustment(study_df = with_ids_dt, col_names = "p_value", level = "chromosome", test = "fdr")
}
\description{
the column name is p_adj_genome.
}

