% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiple-correction.r
\name{p_adjustment_chrwide}
\alias{p_adjustment_chrwide}
\title{This function add a new column to the datatable created by bonferroni correction
the column name is p_adj_genome. COLUMN required cmh_p_val}
\usage{
p_adjustment_chrwide(in_un_adj_p_val_snps_data_file_path,
  out_file_path = NULL, col_names = c("cmh_p_val"), p_adj_method = "fdr")
}
\arguments{
\item{SNP}{p value stats data table}
}
\value{
Datable with additional columns for genomewide correction of each column vector
}
\description{
This function add a new column to the datatable created by bonferroni correction
the column name is p_adj_genome. COLUMN required cmh_p_val
}

