% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps-to-genes.r
\name{pgi_execute_map_snp_to_genes}
\alias{pgi_execute_map_snp_to_genes}
\title{pgi_execute_map_snp_to_genes( in_genome_data_file_path = "/home/data/nehil_genome_p_adj_snp_annotated_study.tsv", 
in_ref_gene_id_file_path = "/home/data/reference/77/cattle_gene_list-UMD3.1.77.txt",
out_snps_in_genes_file_path = "/home/data/cattle_gene_level.tsv")}
\usage{
pgi_execute_map_snp_to_genes(in_genome_data_file_path, sep1,
  in_ref_gene_id_file_path, sep2, out_snps_in_genes_file_path, sep3,
  window_size = 1000)
}
\description{
pgi_execute_map_snp_to_genes( in_genome_data_file_path = "/home/data/nehil_genome_p_adj_snp_annotated_study.tsv", 
in_ref_gene_id_file_path = "/home/data/reference/77/cattle_gene_list-UMD3.1.77.txt",
out_snps_in_genes_file_path = "/home/data/cattle_gene_level.tsv")
}

