% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/get-snp-ids.r
\name{combine_files_in_dir}
\alias{combine_files_in_dir}
\title{This function combines the files in the geiven directory
The assumption is that it does not have a header (default)}
\usage{
combine_files_in_dir(dir_path, header = F, col_names = NULL)
}
\description{
This function combines the files in the geiven directory
The assumption is that it does not have a header (default)
}

