% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dir_merge}
\alias{dir_merge}
\title{Combines all the files in a directory using a Full Outer Join `merge(.., all=T)`}
\usage{
dir_merge(dir_path, col_names, sep = "\\t", out_file_path = NULL)
}
\description{
@param dir_path Absolute path of the directory. Absolute path is recommended to prevent errors.
 @param col_names A character vector of column names used in the condition to match rows from different files.
 @param sep A separator string. It is recommended to use tab in gwas studies.
 @param out_file_path Absolute path of for writing the output. This is Optional
 
 @return a data.table of all the combined files
 @example dir_rbind(~/data/", header = F, col_names = c("chr","pos","allele","p_value"))
}

