% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-snp-ids.r
\name{generate_new_ids}
\alias{generate_new_ids}
\title{Get new snp ids for snps not found in ref db.}
\usage{
generate_new_ids(df, prefix_str = "dal_snp", col_name = "snp_id")
}
\arguments{
\item{col_name, }{a character value used as column name of the snp ids}

\item{the}{dataframe df}

\item{prefix, }{a chacracter value used as prefix to numerical index of the snps, default is 'dal_snp'}
}
\value{
a datatable with all missing snp names changed to dal_snp1, dal_snp2,...
}
\description{
The function takes in the data frame after attaching snp ids from the ref. 
using get_snp_ids(). The function computes the length of missing snp_name
}

