% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-snp-ids.r
\name{get_snp_ids}
\alias{get_snp_ids}
\title{To get the Snp IDs of the snps found in the study. Using columns chr_no, snp_pos, ref_allele, in alt_allele 
using these conditions rows are joined between the two dataframes provided to this function}
\usage{
get_snp_ids(df1, ref_df, out_file_path = NULL)
}
\arguments{
\item{ref_df}{A csv refernce file path (absolute)}

\item{data_path}{A csv file path (absolute) with data about snps and stats from different studies}
}
\value{
returns a data table with all the ref. snp ids joined to each row
}
\description{
To get the Snp IDs of the snps found in the study. Using columns chr_no, snp_pos, ref_allele, in alt_allele 
using these conditions rows are joined between the two dataframes provided to this function
}

