% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/significant-snps.r
\name{get_topQ}
\alias{get_topQ}
\title{Internally calls another function get_quartile}
\usage{
get_topQ(df, column_name, out_file_path = NULL)
}
\description{
Internally calls another function get_quartile
}

