% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snps-to-genes.r
\name{map_snps_to_gene}
\alias{map_snps_to_gene}
\title{The function gets genome data.table and reference data table
It finds all the snps in genome that are in  gene +/- window_size}
\usage{
map_snps_to_gene(genome_dt, ref_dt, window_size)
}
\arguments{
\item{genome_dt}{this is a datatable after bonferroni correction. Should have column names, chr_no, pos}

\item{ref_dt}{this is the Bos taraus dt which has gene_start, gene_end and chr_no column names}

\item{window_size}{}
}
\value{
a dt with all snps in genes
}
\description{
The function gets genome data.table and reference data table
It finds all the snps in genome that are in  gene +/- window_size
}

