% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/get-snp-ids.r
\name{norm_var_names}
\alias{norm_var_names}
\title{This function normalises the input string vector
_ and small case output, no spaces, no . etc}
\usage{
norm_var_names(vars, sep = "_")
}
\arguments{
\item{a}{character vector of names}
}
\value{
a normalised character vector of names
}
\description{
This function normalises the input string vector
_ and small case output, no spaces, no . etc
}

