% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiple-correction.r
\name{p_adjustment_genomewide}
\alias{p_adjustment_genomewide}
\title{This function add a new column to the datatable created by bonferroni correction
should have a column 'chr' which is integers only}
\usage{
p_adjustment_genomewide(in_un_adj_p_val_snps_data_file_path, col_names,
  out_file_path = NULL)
}
\arguments{
\item{in_un_adj_p_val_snps_data_file_path}{is tab separated filepath. This is the file with
unadjusted p valu columns.}

\item{col_names}{Vector of columns to be acted upon}

\item{out_file_path}{Optional parameter. The file path to store the output.}
}
\value{
Datable with additional columns for genomewide correction of each column vector
}
\description{
the column name is p_adj_genome.
}

