% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiple-correction.r
\name{p_adjustment_summary}
\alias{p_adjustment_summary}
\title{column names in order chromosome number, snp id, snp position or base pair, pvalue and adjusted pvalue}
\usage{
p_adjustment_summary(df, col_names)
}
\description{
column names in order chromosome number, snp id, snp position or base pair, pvalue and adjusted pvalue
}

